Remove droplet radius limit in no-ice model
Michael and I realised that the restriction on the cloud-droplet radius
was causing the number of cloud-droplets to increase nonsensically.
Since cloud-water gets removed by the creation of rain-droplets and so
that amount of cloud water decreases in time, it was decided that the
droplet radius was allowed to increase.